% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpt.R
\name{announce_calibration}
\alias{announce_calibration}
\title{Message the user about calibrated bounds}
\usage{
announce_calibration(rb, n_pairs)
}
\arguments{
\item{rb}{Output of \code{\link{resolve_bounds}}.}

\item{n_pairs}{Number of pre-period pairs used.}
}
\description{
Message the user about calibrated bounds
}
\keyword{internal}
