% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbours.R
\name{find_neighbours}
\alias{find_neighbours}
\title{Find matrix of n_neigh neighbours for each node}
\usage{
find_neighbours(nodes_x, nodes_y, n_neigh)
}
\arguments{
\item{nodes_x}{A numeric vector containing the x-coordinates of the nodes.}

\item{nodes_y}{A numeric vector containing the y-coordinates of the nodes.}

\item{n_neigh}{An integer specifying the number of nearest neighbors to find for each node (excluding itself).}
}
\value{
A matrix where each row corresponds to a node, and the columns represent its n_neigh nearest neighbors.
}
\description{
This function computes the matrix of n_neigh nearest neighbors for each node based on their coordinates (nodes_x and nodes_y).
The result is a matrix where each row corresponds to a node, and the columns represent its n_neigh nearest neighbors.
}
\examples{
nodes_x <- c(1, 2, 3, 4, 5)
nodes_y <- c(2, 4, 6, 8, 10)
n_neigh <- 2
find_neighbours(nodes_x, nodes_y, n_neigh)
}
