% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDW_interpolation_nodes.R
\name{IDW_interpolation_nodes}
\alias{IDW_interpolation_nodes}
\title{Inverse Distance Weighted (IDW) Interpolation for Flow Field Data}
\usage{
IDW_interpolation_nodes(
  x,
  y,
  nodes_x,
  nodes_y,
  vel_x,
  vel_y,
  K,
  id.node,
  lst_neighbours
)
}
\arguments{
\item{x}{The x-coordinate of the point where the interpolation is performed.}

\item{y}{The y-coordinate of the point where the interpolation is performed.}

\item{nodes_x}{A numeric vector containing the x-coordinates of the neighboring nodes.}

\item{nodes_y}{A numeric vector containing the y-coordinates of the neighboring nodes.}

\item{vel_x}{A numeric vector containing the x-component of flow velocities at the neighboring nodes.}

\item{vel_y}{A numeric vector containing the y-component of flow velocities at the neighboring nodes.}

\item{K}{A numeric vector containing the turbulent kinematic properties (K) at the neighboring nodes.}

\item{id.node}{The index of the node for which the interpolation is being performed.}

\item{lst_neighbours}{A matrix representing the matrix of nearest neighbors for each node in the flow field.}
}
\value{
A data frame containing the interpolated flow velocities (vel_x and vel_y) and the turbulent kinematic property (int_K) at the given point (x, y). The data frame also includes the index of the closest node (node.id) used for the interpolation.
}
\description{
This function performs Inverse Distance Weighted (IDW) interpolation to estimate the flow velocities (x and y components) and turbulent kinematic properties (K) at a given point (x, y) in a flow field based on the values at neighboring nodes. IDW is a simple spatial interpolation method that gives higher weight to closer neighboring nodes, resulting in a smooth interpolated value.
}
\examples{
# Sample flow field data
nodes_x <- c(1, 2, 3, 4, 5)
nodes_y <- c(2, 4, 6, 8, 10)
vel_x <- c(0.1, 0.2, 0.3, 0.4, 0.5)
vel_y <- c(0.5, 0.4, 0.3, 0.2, 0.1)
K <- c(0.01, 0.02, 0.03, 0.04, 0.05)
lst_neighbours <- matrix(c(2, 3, 1, 3, 4, 2, 4, 5, 3, 5, 4, 2, 5, 3, 1), ncol = 3)
IDW_interpolation_nodes(2.5, 5.5, nodes_x, nodes_y, vel_x, vel_y, K, 3, lst_neighbours)
}
