\name{LPT-package}
\alias{LPT-package}
\alias{LPT}
\docType{package}
\title{
  \packageTitle{LPT}
}
\description{
  \packageDescription{LPT}
}
\details{
  The DESCRIPTION file:
  \packageDESCRIPTION{LPT}
  \packageIndices{LPT}
  The package includes several functions which implements two-sample permutation tests.
}
\author{
  \packageAuthor{LPT}
  
  Maintainer: \packageMaintainer{LPT}
}
\keyword{ package }
\examples{
  x <- rnorm(5)
  y <- rnorm(10)
  
  L1.test(x, y)
  L2.test(x, y)
}
