% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.lpt}
\alias{print.lpt}
\title{Print an lpt object}
\usage{
\method{print}{lpt}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"lpt"}.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print an lpt object
}
