% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.lpt}
\alias{summary.lpt}
\title{Summarize an lpt object}
\usage{
\method{summary}{lpt}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"lpt"}.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{object}, invisibly.
}
\description{
Summarize an lpt object
}
