% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpt.R
\name{wants_calibration}
\alias{wants_calibration}
\title{Does the M/B specification require pre-period calibration?}
\usage{
wants_calibration(M, B, has_untreated)
}
\arguments{
\item{M, B}{User-supplied specifications.}

\item{has_untreated}{Logical. Calibrating M requires untreated units.}
}
\description{
Does the M/B specification require pre-period calibration?
}
\keyword{internal}
