% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_timestep.R
\name{adaptive_timestep}
\alias{adaptive_timestep}
\title{Compute Adaptive Time Step at Each Iteration}
\usage{
adaptive_timestep(distance, vel, vel_min = 0.1)
}
\arguments{
\item{distance}{The distance between the current position and the next position in the simulation.}

\item{vel}{The flow velocity at the current position.}
}
\value{
The computed adaptive time step for the simulation.
}
\description{
This function calculates an adaptive time step to use in a simulation based on the distance to the next position and the current flow velocity. The adaptive time step helps improve the accuracy and stability of the simulation by adjusting the time step dynamically based on the current flow conditions.
}
\examples{
# Calculate the adaptive time step for a simulation
distance <- 2.5 # meters
vel <- 0.1 # meters per second
adaptive_timestep(distance, vel)
}
