% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPT_funs.R
\name{LPT}
\alias{LPT}
\title{Applies the locally most powerful kernel-based transformation to a vector of quantitative responses.}
\usage{
LPT(Y)
}
\arguments{
\item{Y}{a numeric vectors of quantitative responses (or residuals).}
}
\value{
The transformed responses.
}
\description{
Applies the locally most powerful kernel-based transformation to a vector of quantitative responses.
}
\examples{
An example about how to apply the transformation to variant set analysis in genetic association studies.
library(MORST) # For set-based tests
# Generate random phenotype data under null hypothesis
set.seed(123)
n = 1000 ## number of individuals
p = 20 ## number of SNPs
# obtain simulated genotype data
G <- matrix(sample(0:2, n * p, replace = TRUE),nrow = n, ncol = p)
Z <- rnorm(n) # covariates
# Generate standard normal phenotype under H0
Y = Z*1.2+rnorm(n)
Y_res = lm(Y~Z)$residuals
# Transform phenotype using the proposed LPT method
Y_trans = LPT(Y_res)
# Fit null model (without genetic variants)
nullmodel = Null_model_glm(Y_trans,Z)
# Example of runing set-based association tests
# SetBasedTests will give Burden, SKAT, ACAT, MORST results
# weights.beta controls the weighting of rare vs. common variants
SetBasedTests(G,obj = nullmodel,weights.beta = c(1,1))
}
\references{
Liu, Y. and Wang, T. (2025+). A powerful transformation of quantitative responses for biobank-scale association studies.  \emph{Journal of the American Statistical Association}, To apprear.
}
