% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trajectory.R
\name{simulate_trajectory}
\alias{simulate_trajectory}
\title{Simulate Trajectory with Nearest Neighbor Interpolation}
\usage{
simulate_trajectory(flow_df, pos0, n.it, up, vel_coef = 1, lst_neighbours)
}
\arguments{
\item{flow_df}{A data frame containing flow data, including `x`, `y`, `vel_x`, `vel_y`,
`K`, `depth`, `boundary`, and other relevant columns.}

\item{pos0}{A list or data frame representing the initial position of the larva,
containing `x`, `y`, and other relevant columns.}

\item{n.it}{An integer specifying the number of iterations for simulating the trajectory.}

\item{up}{A numeric value representing the vertical velocity of the larva.}

\item{vel_coef}{A numeric coefficient used to correct the velocity based on position.}

\item{lst_neighbours}{A matrix representing the nearest neighbors for each node in the flow
field. This matrix is used for nearest neighbor interpolation.}
}
\value{
A list containing the simulated trajectory of the larva (`traj_larvae`),
        information about the drift destination (`traj_info`), and the final position
        of the larva (`posf`).
}
\description{
This function simulates the trajectory of a larva in a flow field based on given
initial positions and flow data. The larva's trajectory is computed by updating
its position iteratively according to the flow velocities at each step, until
certain conditions are met (e.g., reaching a "safe zone" or a boundary).
}
\details{
Nearest neighbor interpolation is used to find the flow velocities and kinematic
properties (e.g., turbulence) at each iteration. The interpolation is performed using
the provided `lst_neighbours`, which represents the matrix of nearest neighbors
for each node in the flow field.
}
\examples{
# Define flow data and initial position
flow_df <- data.frame(
  x = c(1, 2, 3, 4, 5),
  y = c(2, 4, 6, 8, 10),
  vel_x = c(0.1, 0.2, 0.3, 0.4, 0.5),
  vel_y = c(0.5, 0.4, 0.3, 0.2, 0.1),
  K = c(0.01, 0.02, 0.03, 0.04, 0.05),
  depth = c(1, 2, 3, 4, 5),
  boundary = c("dry", "wet", "wet", "outflow", "dry")
)
pos0 <- list(x = 1, y = 2, id = 1)
n.it <- 100
up <- 0.1
lst_neighbours <- matrix(c(2, 3, 1, 3, 4, 2, 4, 5, 3, 5, 4, 2, 5, 3, 1), ncol = 3)
simulate_trajectory2(flow_df, pos0, n.it, up, lst_neighbours)
}
